feat: Criando testes de bdd e configurações

+++ src/main/kotlin/br/com/manafood/manafoodproduct/adapter/request/commands/update/UpdateProductRequest.kt
@@ -3,13 +3,14 @@ package br.com.manafood.manafoodproduct.adapter.request.commands.update
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Positive
+import java.math.BigDecimal
 import java.util.*
 
 data class UpdateProductRequest(
     @field:NotNull val id: UUID,
     @field:NotBlank val name: String,
     val description: String?,
-    @field:Positive val unitPrice: Double,
+    @field:Positive val unitPrice: BigDecimal,
     val categoryId: UUID,
     val itemIds: List<UUID> = emptyList()
 )
